Add files via upload
This repository contains a  team project for Myntra Hackerramp 2024 in which we are using news API to get relevant data for trend analysis and forecast. The future scope of this project includes integrating multiple data sources like fashion blogs , twitter API, etc. to get more data for real time analysis and forecast.